mpsc: blocking push_block / pop_block + extract shared park module
Lift the futex-style park/unpark infrastructure out of mpmc into a
new common::park module (WakeSet, PARK_SLOTS, BACKOFF_PARK_THRESHOLD,
PARK_TIMEOUT_MICROS) and reuse it in mpsc to add Producer::push_block
and Consumer::pop_block. Producers share a 64-bit wake bitmap (one
park slot per producer clone, modulo 64); the single consumer parks
on a OnceLock<Thread> + AtomicBool flag. push_block returns Err(val)
when the consumer drops; pop_block returns None only after all
producers drop AND the queue drains.

Wakes are wired into push, drain, drain_up_to, pop, SlotReader::Drop,
WrittenSlot::commit, and SlotWriter::Drop. Consumer::Drop sets
consumer_closed *before* draining so parked producers don't race the
drain to push into freed slots.

Cleanup pass on the audit:
- mpsc::wake_consumer free fn -> RingBuffer::wake_consumer method.
- mpmc scan_unused free fn -> RingBuffer::scan_unused method.
- Remove redundant `if wake.load() != 0 { wake_one() }` gate at 6
  call sites; WakeSet::wake_one is #[inline] and self-gates already.
- Fix all 15 pre-existing clippy warnings without #[allow]: split
  mpmc::Producer::push into refill_batch/park_until_slot_free helpers
  to drop under the 100-line cap, replace u64-as-usize casts with
  usize::try_from, fix doc backticks, add explicit `drop(p)` to keep
  producer alive past push_block test joins.

153/153 tests pass, 0 clippy warnings.

+++ src/broadcast/mod.rs
@@ -968,10 +968,11 @@ mod tests {
     /// the producer cannot reclaim that slot, so a fill-then-push fails.
     #[test]
     fn pop_ref_blocks_producer_overwrite() {
-        let cap = 4;
-        let (producer, mut consumer) = RingBuffer::<u32>::new(Capacity::exact(cap), 4).split();
+        let cap = 4u32;
+        let (producer, mut consumer) =
+            RingBuffer::<u32>::new(Capacity::exact(cap as usize), 4).split();
 
-        for i in 0..cap as u32 {
+        for i in 0..cap {
             producer.push(i).unwrap();
         }
         assert!(producer.push(99).is_err(), "buffer is full");
